requirements.txt
configparser == 5.2.0
dnspython==2.0.0
pymongo == 4.0.2
tweepy == 3.10.0
nltk == 3.7
python-dateutil == 2.8.2
pytest == 6.2.4
mongomock == 4.0.0
mockupdb == 1.8.1
